+++ glfw/src/VideoMode.zig
@@ -10,33 +10,33 @@ const VideoMode = @This();
 handle: c.GLFWvidmode,
 
 /// Returns the width of the video mode, in screen coordinates.
-pub inline fn getWidth(self: VideoMode) usize {
-    return @intCast(usize, self.handle.width);
+pub inline fn getWidth(self: VideoMode) u32 {
+    return @intCast(u32, self.handle.width);
 }
 
 /// Returns the height of the video mode, in screen coordinates.
-pub inline fn getHeight(self: VideoMode) usize {
-    return @intCast(usize, self.handle.height);
+pub inline fn getHeight(self: VideoMode) u32 {
+    return @intCast(u32, self.handle.height);
 }
 
 /// Returns the bit depth of the red channel of the video mode.
-pub inline fn getRedBits(self: VideoMode) usize {
-    return @intCast(usize, self.handle.redBits);
+pub inline fn getRedBits(self: VideoMode) u32 {
+    return @intCast(u32, self.handle.redBits);
 }
 
 /// Returns the bit depth of the green channel of the video mode.
-pub inline fn getGreenBits(self: VideoMode) usize {
-    return @intCast(usize, self.handle.greenBits);
+pub inline fn getGreenBits(self: VideoMode) u32 {
+    return @intCast(u32, self.handle.greenBits);
 }
 
 /// Returns the bit depth of the blue channel of the video mode.
-pub inline fn getBlueBits(self: VideoMode) usize {
-    return @intCast(usize, self.handle.blueBits);
+pub inline fn getBlueBits(self: VideoMode) u32 {
+    return @intCast(u32, self.handle.blueBits);
 }
 
 /// Returns the refresh rate of the video mode, in Hz.
-pub inline fn getRefreshRate(self: VideoMode) usize {
-    return @intCast(usize, self.handle.refreshRate);
+pub inline fn getRefreshRate(self: VideoMode) u32 {
+    return @intCast(u32, self.handle.refreshRate);
 }
 
 test "getters" {
